Version `ucsi` module
Move existing UCSI code under `v1_2` and put it behind `ucsi-v1_2`
feature.
