+++ src/ucsi/v1_2/lpm/connector_reset.rs
@@ -1,20 +1,19 @@
-//! PPM Reset command
-
-use bincode::de::{Decode, Decoder};
-use bincode::enc::{Encode, Encoder};
+//! Types for CONNECTOR_RESET command, see UCSI spec 6.5.3
+use bincode::de::Decoder;
+use bincode::enc::Encoder;
 use bincode::error::{DecodeError, EncodeError};
+use bincode::{Decode, Encode};
 
-use crate::ucsi::{CommandHeaderRaw, COMMAND_LEN};
+use crate::ucsi::v1_2::{CommandHeaderRaw, COMMAND_LEN};
 
-/// PPM Reset command structure
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+/// Command padding
+// -1 for the connector number byte
+pub const COMMAND_PADDING: usize = COMMAND_LEN - size_of::<CommandHeaderRaw>() - 1;
+
+/// Command arguments
+#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct Args;
-
-/// Data length for the PPM_RESET command response
-pub const RESPONSE_DATA_LEN: u8 = 0;
-/// Command padding
-pub const COMMAND_PADDING: usize = COMMAND_LEN - size_of::<CommandHeaderRaw>();
 
 impl Encode for Args {
     fn encode<E: Encoder>(&self, encoder: &mut E) -> Result<(), EncodeError> {
