+++ src/ucsi/v1_2/ppm/get_capability.rs
@@ -5,7 +5,7 @@ use bincode::enc::{Encode, Encoder};
 use bincode::error::{DecodeError, EncodeError};
 use bitfield::bitfield;
 
-use crate::ucsi::{CommandHeaderRaw, COMMAND_LEN};
+use crate::ucsi::v1_2::{CommandHeaderRaw, COMMAND_LEN};
 
 /// Data length for the GET_CAPABILITY command response
 pub const RESPONSE_DATA_LEN: usize = 16;
